@@ -1,8 +1,7 @@
-﻿namespace LibraryManager.Domain.Dtos.Author
+﻿namespace LibraryManager.Domain.Dtos.Author;
+
+public class GetAuthorDto
 {
-    public class GetAuthorDto
-    {
-        public long AuthorId { get; set; }
-        public string Name { get; set; }
-    }
+    public long AuthorId { get; set; }
+    public string Name { get; set; }
 }
