@@ -1,9 +1,8 @@
-﻿namespace LibraryManager.Domain.Dtos.Books
+﻿namespace LibraryManager.Domain.Dtos.Books;
+
+public class LateBookWithStudentNameDto
 {
-    public class LateBookWithStudentNameDto
-    {
-        public long BookId { get; set; }
-        public string BookTitle { get; set; }
-        public string StudentName { get; set; }
-    }
+    public long BookId { get; set; }
+    public string BookTitle { get; set; }
+    public string StudentName { get; set; }
 }
