@@ -1,7 +1,6 @@
-﻿namespace LibraryManager.Domain.Dtos.Books
+﻿namespace LibraryManager.Domain.Dtos.Books;
+
+public class ReturnBookDto
 {
-    public class ReturnBookDto
-    {
-        public long BookId { get; set; }
-    }
+    public long BookId { get; set; }
 }
