@@ -1,9 +1,8 @@
-﻿namespace LibraryManager.Domain.Dtos.Books
+﻿namespace LibraryManager.Domain.Dtos.Books;
+
+public class UpdateBookDto
 {
-    public class UpdateBookDto
-    {
-        public long Id { get; set; }
-        public string Title { get; set; }
-        public string Description { get; set; }
-    }
+    public long Id { get; set; }
+    public string Title { get; set; }
+    public string Description { get; set; }
 }
