@@ -1,9 +1,8 @@
-﻿namespace LibraryManager.Domain.Dtos
+﻿namespace LibraryManager.Domain.Dtos;
+
+public class ChangePasswordDto
 {
-    public class ChangePasswordDto
-    {
-        public string Email { get; set; }
-        public string OldPassword { get; set; }
-        public string NewPassword { get; set; }
-    }
+    public string Email { get; set; }
+    public string OldPassword { get; set; }
+    public string NewPassword { get; set; }
 }
