@@ -1,8 +1,7 @@
-﻿namespace LibraryManager.Domain.Dtos
+﻿namespace LibraryManager.Domain.Dtos;
+
+public class LoginDto
 {
-    public class LoginDto
-    {
-        public string Email { get; set; }
-        public string Password { get; set; }
-    }
+    public string Email { get; set; }
+    public string Password { get; set; }
 }
