@@ -1,8 +1,7 @@
-﻿namespace LibraryManager.Domain.Dtos.Students
+﻿namespace LibraryManager.Domain.Dtos.Students;
+
+public class CreateStudentDto
 {
-    public class CreateStudentDto
-    {
-        public string Name { get; set; }
-        public long TeacherId { get; set; }
-    }
+    public string Name { get; set; }
+    public long TeacherId { get; set; }
 }
