@@ -1,8 +1,7 @@
-﻿namespace LibraryManager.Domain.Dtos.Students
+﻿namespace LibraryManager.Domain.Dtos.Students;
+
+public class GetStudentsDto
 {
-    public class GetStudentsDto
-    {
-        public long StudentId { get; set; }
-        public string Name { get; set; }
-    }
+    public long StudentId { get; set; }
+    public string Name { get; set; }
 }
