@@ -1,9 +1,8 @@
-﻿namespace LibraryManager.Domain.Dtos.Students
+﻿namespace LibraryManager.Domain.Dtos.Students;
+
+public class StudentsWithBooksDto
 {
-    public class StudentsWithBooksDto
-    {
-        public long StudentId { get; set; }
-        public string Name { get; set; }
-        public int NumberOfActiveBooks { get; set; }
-    }
+    public long StudentId { get; set; }
+    public string Name { get; set; }
+    public int NumberOfActiveBooks { get; set; }
 }
