@@ -1,13 +1,10 @@
-﻿namespace LibraryManager.Domain.Entities
+﻿namespace LibraryManager.Domain.Entities;
+
+public class BaseEntity
 {
-    using System;
+    public long Id { get; set; }
 
-    public class BaseEntity
-    {
-        public long Id { get; set; }
+    public DateTimeOffset CreateDate { get; set; }
 
-        public DateTimeOffset CreateDate { get; set; }
-
-        public bool IsRemoved { get; set; }
-    }
+    public bool IsRemoved { get; set; }
 }
