@@ -1,34 +1,29 @@
-﻿namespace LibraryManager.Domain.Entities.Configurations
+﻿namespace LibraryManager.Domain.Entities.Configurations;
+
+public sealed class TransactionEntityConfiguration : IEntityTypeConfiguration<Transaction>
 {
-    using Microsoft.EntityFrameworkCore;
-    using Microsoft.EntityFrameworkCore.Metadata.Builders;
-    using System;
-
-    public sealed class TransactionEntityConfiguration : IEntityTypeConfiguration<Transaction>
+    public void Configure(EntityTypeBuilder<Transaction> builder)
     {
-        public void Configure(EntityTypeBuilder<Transaction> builder)
-        {
-            builder.ToTable("Transaction")
-                .HasKey(x => x.Id);
+        builder.ToTable("Transaction")
+            .HasKey(x => x.Id);
 
-            builder.Property(x => x.Id).ValueGeneratedOnAdd();
-            builder.Property(x => x.BookId);
-            builder.Property(x => x.StudentId);
-            builder.Property(x => x.CreateDate)
-                .IsRequired()
-                .HasDefaultValue(DateTimeOffset.UtcNow);
-            builder.Property(x => x.LendDate);
-            builder.Property(x => x.IsRemoved)
-                .IsRequired()
-                .HasDefaultValue(false);
+        builder.Property(x => x.Id).ValueGeneratedOnAdd();
+        builder.Property(x => x.BookId);
+        builder.Property(x => x.StudentId);
+        builder.Property(x => x.CreateDate)
+            .IsRequired()
+            .HasDefaultValue(DateTimeOffset.UtcNow);
+        builder.Property(x => x.LendDate);
+        builder.Property(x => x.IsRemoved)
+            .IsRequired()
+            .HasDefaultValue(false);
 
-            builder.HasOne(x => x.Student)
-                .WithMany(x => x.Transactions)
-                .OnDelete(DeleteBehavior.SetNull);
+        builder.HasOne(x => x.Student)
+            .WithMany(x => x.Transactions)
+            .OnDelete(DeleteBehavior.SetNull);
 
-            builder.HasOne(x => x.Book)
-                .WithMany(x => x.Transactions)
-                .OnDelete(DeleteBehavior.SetNull);
-        }
+        builder.HasOne(x => x.Book)
+            .WithMany(x => x.Transactions)
+            .OnDelete(DeleteBehavior.SetNull);
     }
 }
