@@ -1,23 +1,20 @@
-﻿namespace LibraryManager.Domain.Entities
+﻿namespace LibraryManager.Domain.Entities;
+
+public sealed class Transaction : BaseEntity
 {
-    using System;
+    public bool Active { get; set; } = true;
 
-    public sealed class Transaction : BaseEntity
-    {
-        public bool Active { get; set; } = true;
+    public Book? Book { get; set; }
 
-        public Book? Book { get; set; }
+    public long? BookId { get; set; }
 
-        public long? BookId { get; set; }
+    public DateTimeOffset LendDate { get; set; }
 
-        public DateTimeOffset LendDate { get; set; }
+    public Student? Student { get; set; }
 
-        public Student? Student { get; set; }
+    public long? StudentId { get; set; }
 
-        public long? StudentId { get; set; }
+    public DateTimeOffset? ReturnedAt { get; set; }
 
-        public DateTimeOffset? ReturnedAt { get; set; }
-
-        public DateTimeOffset ReturnDate { get; set; }
-    }
+    public DateTimeOffset ReturnDate { get; set; }
 }
