@@ -1,9 +1,10 @@
-﻿namespace LibraryManager.Domain.Entities
+﻿namespace LibraryManager.Domain.Entities;
+
+public class User : BaseEntity
 {
-    public class User : BaseEntity
-    {
-        public string Email { get; set; }
-        public string Password { get; set; }
-        public string Role { get; set; }
-    }
+    public string Email { get; set; }
+
+    public string Password { get; set; }
+
+    public string Role { get; set; }
 }
