@@ -1,11 +1,7 @@
-﻿using LibraryManager.Domain.Dtos;
-using System.Threading.Tasks;
+﻿namespace LibraryManager.Domain.Abstractions.Services;
 
-namespace LibraryManager.Domain.Abstractions.Services
+public interface IAuthService
 {
-    public interface IAuthService
-    {
-        Task<bool> Register(RegisterDto registerDto);
-        Task<bool> ChangePassword(ChangePasswordDto changePasswordDto);
-    }
+    Task<bool> Register(RegisterDto registerDto);
+    Task<bool> ChangePassword(ChangePasswordDto changePasswordDto);
 }
