@@ -1,17 +1,10 @@
-﻿using LibraryManager.Domain.Dtos.Books;
-using LibraryManager.Domain.Dtos.Transactions;
-using LibraryManager.Domain.Entities;
-using System.Collections.Generic;
-using System.Threading.Tasks;
+﻿namespace LibraryManager.Domain.Abstractions.Services;
 
-namespace LibraryManager.Domain.Abstractions.Services
+public interface ITransactionService
 {
-    public interface ITransactionService
-    {
-        Task<IEnumerable<Transaction>> GetAll();
-        Task<IEnumerable<Transaction>> GetAllByBook(long bookId);
-        Task<IEnumerable<LateBookWithStudentNameDto>> GetLateBooksWithStudentName(long teacherId);
-        Task<IEnumerable<Transaction>> GetAllActiveTransactions();
-        Task<IEnumerable<GetTransactionDto>> GetTransactionsWithDetailsByStudent(long studentId);
-    }
+    Task<IEnumerable<Transaction>> GetAll();
+    Task<IEnumerable<Transaction>> GetAllByBook(long bookId);
+    Task<IEnumerable<LateBookWithStudentNameDto>> GetLateBooksWithStudentName(long teacherId);
+    Task<IEnumerable<Transaction>> GetAllActiveTransactions();
+    Task<IEnumerable<GetTransactionDto>> GetTransactionsWithDetailsByStudent(long studentId);
 }
