@@ -1,11 +1,8 @@
-﻿using System.ComponentModel.DataAnnotations;
+﻿namespace LibraryManager.Domain.Dtos.Author;
 
-namespace LibraryManager.Domain.Dtos.Author
+public class CreateAuthorDto
 {
-    public class CreateAuthorDto
-    {
-        [Required]
-        [StringLength(100, MinimumLength = 1)]
-        public string Name { get; set; }
-    }
+    [Required]
+    [StringLength(100, MinimumLength = 1)]
+    public string Name { get; set; }
 }
