@@ -1,14 +1,11 @@
-﻿using System.ComponentModel.DataAnnotations;
+﻿namespace LibraryManager.Domain.Dtos;
 
-namespace LibraryManager.Domain.Dtos
+public class CreateBookDto
 {
-    public class CreateBookDto
-    {
-        [Required]
-        public int AuthorId { get; set; }
-        public string Description { get; set; }
-        public int Reference { get; set; }
-        [Required]
-        public string Title { get; set; }
-    }
+    [Required]
+    public int AuthorId { get; set; }
+    public string Description { get; set; }
+    public int Reference { get; set; }
+    [Required]
+    public string Title { get; set; }
 }
