@@ -1,13 +1,9 @@
-﻿using LibraryManager.Domain.Dtos.Transactions;
-using System.Collections.Generic;
+﻿namespace LibraryManager.Domain.Dtos.Books;
 
-namespace LibraryManager.Domain.Dtos.Books
+public class GetBookDto
 {
-    public class GetBookDto
-    {
-        public long BookId { get; set; }
-        public string Title { get; set; }
-        public string Description { get; set; }
-        public IEnumerable<GetTransactionDto> Transactions { get; set; }
-    }
+    public long BookId { get; set; }
+    public string Title { get; set; }
+    public string Description { get; set; }
+    public IEnumerable<GetTransactionDto> Transactions { get; set; }
 }
