@@ -1,19 +1,16 @@
-﻿using System.ComponentModel.DataAnnotations;
+﻿namespace LibraryManager.Domain.Dtos.Teacher;
 
-namespace LibraryManager.Domain.Dtos.Teacher
+public class UpdateTeacherDto
 {
-    public class UpdateTeacherDto
-    {
-        [Required]
-        public long Id { get; set; }
+    [Required]
+    public long Id { get; set; }
 
-        [Required]
-        public string Name { get; set; }
+    [Required]
+    public string Name { get; set; }
 
-        [Required]
-        public string Email { get; set; }
+    [Required]
+    public string Email { get; set; }
 
-        [Required]
-        public string Role { get; set; }
-    }
+    [Required]
+    public string Role { get; set; }
 }
